Appreciate UIImage.scale when decompressing images
Change `targetSize` to use the same scale as `UIImage.size`. Also use `round` (to nearest) to convert scaled dimensions to integer rather than truncating the fractional part.

Since sizes in iOS are expressed in logical points, not pixels, it is more natural to treat `targetSize` as logical points, so that decompressing an image of `image.size == CGSize(width: 400, height: 300)` and `image.scale == 2.0` down to `targetSize == CGSize(width: 200, height: 150)` would actually mean the same as `decompressImage(image, 0.5)` – rather than `decompressImage(image, 0.25)` which it did before this patch.